pyproject.toml
[tool.poetry]
name = "syuwa_daijiten"
version = "0.1.0"
description = "extract data from 手話・日本語大辞典"
authors = ["Nathan Glenn <nathan-gary.glenn@trivago.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
transformers = "^4.15.0"
Pillow = "^9.0.0"
pdf2image = "^1.16.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
